rewrite calculator into a library
